REFACTOR(Web proj): structure change, because more page specific view models and their serving types are needed
