+++ OnlineMinion.Web/Components/PaymentSpecsEditor.razor
@@ -1,6 +1,6 @@
 @using OnlineMinion.Contracts.PaymentSpec.Responses
 @using OnlineMinion.Contracts.CurrencyInfo.Responses
-@using OnlineMinion.Web.ViewModels.CurrencyInfo
+@using OnlineMinion.Web.CurrencyInfo.ViewModels
 
 @code {
