+++ OnlineMinion.Web/Components/TransactionCreditsEditor.razor
@@ -1,6 +1,5 @@
 @using OnlineMinion.Contracts.Transactions
-@using OnlineMinion.Contracts.PaymentSpec.Responses
-@using OnlineMinion.Web.ViewModels.Transaction.Credit
+@using OnlineMinion.Web.Transaction.Credit.ViewModels
 @using System.Globalization
 
 @code {
