+++ OnlineMinion.Web/Pages/PaymentSpecsPage.razor
@@ -1,5 +1,5 @@
 @page "/payment-specs"
-@using OnlineMinion.Web.ViewModels.CurrencyInfo
+@using OnlineMinion.Web.CurrencyInfo.ViewModels
 @using OnlineMinion.Contracts.PaymentSpec.Responses
 @inherits BaseCRUDPage<
             PaymentSpecResp,
