+++ OnlineMinion.Web/Pages/TransactionCreditsPage.razor
@@ -1,7 +1,7 @@
 @page "/transactions/credit"
 @using static Common.Utilities.TypeMemberHelpers;
-@using static ViewModels.CurrencyInfo.CurrencyInfoExtensions;
-@using OnlineMinion.Web.ViewModels.Transaction.Credit
+@using static CurrencyInfo.ViewModels.CurrencyInfoExtensions;
+@using OnlineMinion.Web.Transaction.Credit.ViewModels
 @inherits BaseCRUDPage<
             TransactionCreditListItem,
             OnlineMinion.Contracts.Transactions.Credit.Responses.TransactionCreditResp,
