+++ OnlineMinion.Web/Pages/TransactionDebitsPage.razor
@@ -1,7 +1,7 @@
 @page "/transactions/debit"
 @using static Common.Utilities.TypeMemberHelpers;
 @using static ViewModels.CurrencyInfo.CurrencyInfoExtensions;
-@using OnlineMinion.Web.ViewModels.Transaction.Debit
+@using OnlineMinion.Web.Transaction.Debit.ViewModels
 @inherits BaseCRUDPage<
             TransactionDebitListItem,
             OnlineMinion.Contracts.Transactions.Debit.Responses.TransactionDebitResp,
